CFG:fix path CPP:fix type
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@4648 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/packetqueue.cpp
@@ -37,8 +37,8 @@ bool packetqueue::flushtolog(const char *logfile)
 
         fputs("ENET PACKET\n", f);
         fprintf(f, "flags == %d\n", p->flags);
-        fprintf(f, "referenceCount == %d\n", p->referenceCount);
-        fprintf(f, "dataLength == %d\n", p->dataLength);
+        fprintf(f, "referenceCount == %d\n", (int)p->referenceCount);
+        fprintf(f, "dataLength == %d\n", (int)p->dataLength);
         fputs("data == \n", f);
         // print whole buffer char-wise
         loopj(p->dataLength)
